doc/examples/PinMonitor/python/requirements.txt
cffi==1.5.0
cryptography==1.2.1
enum34==1.1.2
idna==2.0
ipaddress==1.0.16
ndg-httpsclient==0.4.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
six==1.10.0
wheel==0.26.0

pyserial
